Rearrange file content
