+++ src/angular/parent.ts
@@ -1,22 +1,31 @@
-import * as path from "path";
 import * as vscode from "vscode";
+import { getConfig } from "../config";
 
 /**
- * Insert the selected template code into the template of the child component
- * @param directory Directory of the child component
- * @param componentName Name of the child component
- * @param content Selected content of the parent component
+ * Adjust the template of the parent component
+ * @param componentName Name of the component
+ * @param extensionId Extension ID to get the config
  */
-export const writeChildTemplate = (
-  directory: string,
+export const adjustParentComponentTemplate = (
   componentName: string,
-  content: string
+  extensionId: string
 ) => {
-  const componentPath = path.join(
-    directory,
-    componentName,
-    `${componentName}.component.html`
-  );
-  const componentUri = vscode.Uri.file(componentPath);
-  vscode.workspace.fs.writeFile(componentUri, Buffer.from(content));
+  // Get the active text editor
+  const editor = vscode.window.activeTextEditor;
+
+  const document = editor?.document;
+  const selection = editor?.selection;
+
+  if (!editor || !document || !selection) {
+    throw Error("Something went wrong");
+  }
+
+  // Get the default prefix from the VS Code settings
+  const defaultPrefix = getConfig(extensionId, "default-prefix");
+
+  // Component HTML tags
+  const component = `<${defaultPrefix}-${componentName}></${defaultPrefix}-${componentName}>`;
+  editor.edit((editBuilder) => {
+    editBuilder.replace(selection, component);
+  });
 };
